Checkpointing.
By using `six` I think I have most of the imports squared away.  There's
probably still uses of `unicode` built-ins that need fixing.

The idea is to first get the test suite running (which it doesn't yet), and
then to fix tests.

There's a bug in lazr.config which requires us to patch it for now.