Added IMember.subscriber to definitively return how a member is subscribed to the mailing list (via preferred address/user or explicit address).
IMember.get_member() is defined to return the explicit address when members
are subscribed in both ways.

IMember.get_memberships() returns a sequence of length 0, 1, or 2 containing
all the member records associated with the email address.

Fixed the AbstractMemberRoster methods query to properly return subscriptions
via the user's preferred address and via an explicit address.